services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
redis==5.0.1
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-consul==1.1.0
